refactor: satisfy spa lint budgets and whoop lint hygiene
Pre-push lint surfaced three max-lines/max-lines-per-function errors and
seven whoop test warnings. The connections catalog splits its flow lists
into connection-flows.ts, the chat prefill moves into a useChatPrefill
hook consumed by ChatWorkspace, the push op binding moves into
usePushToGarminOp, and whoop test assertions use named expected values.
