+++ packages/workout-spa-editor/src/hooks/use-chat-action-ops.ts
@@ -16,9 +16,7 @@ import type {
   CreateWorkoutInput,
   LogHealthMetricInput,
   LogIntakeInput,
-  PushToGarminInput,
 } from "../application/chat/tools/chat-tool-deps";
-import { useGarminBridge } from "../contexts";
 import { usePersistence } from "../contexts/persistence-context";
 import type { LlmProviderConfig } from "../store/ai-store-types";
 import {
@@ -27,7 +25,7 @@ import {
   doSyncCoaching,
 } from "./chat/chat-action-ops-impl";
 import { doLogIntake } from "./chat/do-log-intake";
-import { doPushToGarmin } from "./chat/do-push-to-garmin";
+import { usePushToGarminOp } from "./use-push-to-garmin-op";
 
 const requireProfile = (profileId: string | null): string => {
   if (!profileId) throw new Error("No active profile");
@@ -44,7 +42,7 @@ export const useChatActionOps = (
   generation: ChatGenerationModel
 ): ChatActionOps => {
   const persistence = usePersistence();
-  const { pushWorkout } = useGarminBridge();
+  const pushToGarmin = usePushToGarminOp();
   const transport = useMemo(
     () =>
       createTrain2GoCoachingTransport(
@@ -80,11 +78,6 @@ export const useChatActionOps = (
     (input: LogIntakeInput) =>
       doLogIntake(persistence, requireProfile(profileId), input),
     [persistence, profileId]
-  );
-  const pushToGarmin = useCallback(
-    (input: PushToGarminInput) =>
-      doPushToGarmin(persistence, pushWorkout, input.workoutId),
-    [persistence, pushWorkout]
   );
 
   return useMemo(
